ml/requirements.txt
streamlit>=1.24.0
torch>=2.0.0
transformers>=4.30.0
Pillow>=9.5.0
PyMuPDF>=1.22.0
pytesseract>=0.3.10
numpy>=1.24.0 